requirements.txt
discord.py
aiohttp
asyncio
datetime